--- src/main/java/lotto/utils/Converter.java
@@ -1,7 +1,0 @@
-package lotto.utils;
-
-public class Converter {
-    public static int changeUnit(int totalPrice) {
-        return totalPrice / 1000;
-    }
-}
